Drop casts

+++ spring-aop/src/test/java/org/springframework/aop/aspectj/AspectJExpressionPointcutTests.java
@@ -210,11 +210,11 @@ public class AspectJExpressionPointcutTests {
 		CallCountingInterceptor interceptor = new CallCountingInterceptor();
 		TestBean testBean = getAdvisedProxy(expression, interceptor);
 
-		assertThat(interceptor.getCount()).as("Calls should be 0").isEqualTo((long) 0);
+		assertThat(interceptor.getCount()).as("Calls should be 0").isEqualTo(0);
 		testBean.getAge();
-		assertThat(interceptor.getCount()).as("Calls should be 1").isEqualTo((long) 1);
+		assertThat(interceptor.getCount()).as("Calls should be 1").isEqualTo(1);
 		testBean.setAge(90);
-		assertThat(interceptor.getCount()).as("Calls should still be 1").isEqualTo((long) 1);
+		assertThat(interceptor.getCount()).as("Calls should still be 1").isEqualTo(1);
 	}
 
 	@Test
@@ -223,12 +223,12 @@ public class AspectJExpressionPointcutTests {
 		CallCountingInterceptor interceptor = new CallCountingInterceptor();
 		TestBean testBean = getAdvisedProxy(expression, interceptor);
 
-		assertThat(interceptor.getCount()).as("Calls should be 0").isEqualTo((long) 0);
+		assertThat(interceptor.getCount()).as("Calls should be 0").isEqualTo(0);
 		testBean.setSomeNumber(new Double(30));
-		assertThat(interceptor.getCount()).as("Calls should be 1").isEqualTo((long) 1);
+		assertThat(interceptor.getCount()).as("Calls should be 1").isEqualTo(1);
 
 		testBean.setSomeNumber(new Integer(90));
-		assertThat(interceptor.getCount()).as("Calls should be 1").isEqualTo((long) 1);
+		assertThat(interceptor.getCount()).as("Calls should be 1").isEqualTo(1);
 	}
 
 	@Test
